New voms-verify standalone command
The voms-verify command loads a certificate chain from the
X509_USER_CERT env variable and runs verification on it.

The command exits with 0 in case of verification success, with 1 in case
of verification error and with -1 in case of other errors.

This command is used as the base for a VOMS verification testing library.