Allow to skip OpenSSL initialization
Add a class static function to allow clients to skip OpenSSL
initialization. This makes sense only for versions up to 1.0.2; above
that the initialization is done automatically. There is no check on
this though, to minimize changes.

Moreover OpenSSL (if not skipped) and AC- and Proxy-related
functionality are initialized atomically.

The flag that tells if OpenSSL needs to be initialized is not atomic
because it is checked inside the critical section of pthread_once.

Fixes: #89