Declare fields of AC_TARGET as optional
Only one of name, group and cert is allowed.
According to RFC 3281, the AC target should be a choice between the
three fields, but apparently VOMS traditionally implements it as three
optional fields.
The change fixes the parsing on the client side. How the server behaves
needs to be checked.

Fix #102