If a detailed error message is available, do not overwrite
When verification of ACs fails, the prior behavior is to always have
this message:

```
Cannot verify AC signature!
```

This can be difficult to debug as there's no indication of whether
its a problem with the proxy itself or with the host configuration.

This patch appends the underlying error message if one was provided.
For example,

```
Cannot verify AC signature!  Underlying error: Certificate verification \
  failed for certificate '/CN=voms.example.com': certificate has expired.
```

(newlines added for readability)