Revert "Preserve the possible failure of AC verification"
This reverts commit 34bc9cbe978792f71fae8ba4dc42e1381cf248d3.

vomsdata::Retrieve() retrieves information about the AC only if the
verification (if requested) succeeds. voms-proxy-info instead always
prints the AC information, but to do so, if there is a verification
failure, it needs to call vomsdata::Retrieve() a second time, without
asking for verification. The reverted commit didn't do this, so let's
start over.