Make sure to print the AC verification failure message
Also make sure to return failure from the test_proxy function if the AC
verification is requested, i.e. if -dont-verify-ac is not specified,
and the verification fails. This causes voms-proxy-info to also return
failure in such a case, as it was always supposed to do.

Although returning failure is a bug fix, it is in practice also a
backwards incompatibility and an incompatibility with the Java version
of this utility, which doesn't verify the AC at all.