Don't contact any more the legacy endpoint
For many years the VOMS server has offered the REST endpoint
(/generate-ac), which is also the only endpoint offered by VOMS-AA, the
VOMS component of INDIGO-IAM. It's time to retire the legacy endpoint,
based on XML messages.

This change only affects the client side; its main benefit is to
simplify how errors are reported to the user.