Ignore the HTTP status when processing VOMS response
In principle the contents of the VOMS response is meaningful only
when the HTTP status code is 200. However, for reasons at the moment
unknown, the VOMS server sometimes returns "500 Internal Server Error"
despite providing a valid content in the body. This has been seen in
the test "See if requesting a too long proxy fails", which issues
`voms-proxy-init --voms vo.0 --valid 100:00`. The VOMS server log,
even in debug mode, doesn't show anything suspicious.

The fix consists in parsing the body without looking at the HTTP
status code. Note that before, this check was done *after* the parsing,
which doesn't seem totally correct.